WIP Groups are working
Recursive groups work, in terms of loading from a vZome file.  The group
does not function correctly afterward.

There is no support for legacy groups yet.
