+++ client/src/bundles/jsweet.js
@@ -11,6 +11,7 @@ import * as designs from './models'
 import { meshChanged } from './mesh'
 import * as shapers from './shapers'
 import { ActionCreators } from 'redux-undo';
+import Adapter from './adapter'
 
 
 // I can't use the ES6 module approach until I figure out how to use J4TS that way.
@@ -479,155 +480,6 @@ const shaperFactory = ( vzomePkg, orbitSource ) => shapes => instance =>
   const finalColor = rm.getColor().getRGB()
 
   return { id, position: [ x, y, z ], rotation, color: finalColor, shapeId }
-}
-
-class Adapter
-{
-  constructor( shown, selected, hidden )
-  {
-    this.shown = shown
-    this.hidden = hidden
-    this.selected = selected
-  }
-
-  clone()
-  {
-    return new Adapter( new Map( this.shown ), new Map( this.selected ), new Map( this.hidden ) )
-  }
-
-  selectAll()
-  {
-    const temp = this.selected
-    this.selected = this.shown
-    this.shown = temp
-  }
-
-  clearSelection()
-  {
-    const temp = this.shown
-    this.shown = this.selected
-    this.selected = temp
-  }
-
-  delete( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    this.shown.delete( id )
-    this.selected.delete( id )
-    this.hidden.delete( id )
-  }
-
-  select( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    if ( this.selected.has( id ) )
-      return // idempotent
-    const instance = this.shown.get( id )
-    if ( ! instance ) {
-      const hidden = this.hidden.get( id )
-      throw new Error( `No shown instance to select at ${id}`)
-    }
-    this.shown.delete( id )
-    this.selected.set( id, instance )
-  }
-
-  unselect( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    const instance = this.selected.get( id )
-    this.selected.delete( id )
-    this.shown.set( id, instance )
-  }
-
-  selectionSize()
-  {
-    return this.selected.size
-  }
-
-  selectedIterator()
-  {
-    return Array.from( this.selected.values() ).map( instance => instance.vectors ).values()
-  }
-
-  manifestationRendered( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    return this.shown.has( id ) || this.selected.has( id )
-  }
-
-  manifestationHidden( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    return this.hidden.has( id )
-  }
-
-  manifestationSelected( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    return this.selected.has( id )
-  }
-
-  manifestationHasColor( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    const existing = this.shown.get( id ) || this.selected.get( id )
-    return existing && !!existing.color
-  }
-
-  manifestationColor( vectors )
-  {
-    const { id } = mesh.createInstance( vectors )
-    const existing = this.shown.get( id ) || this.selected.get( id )
-    return existing.color
-  }
-
-  setManifestationColor( vectors, color )
-  {
-    const instance = mesh.createInstance( vectors )
-    instance.color = color
-    if ( this.shown.has( instance.id ) ) {
-      this.shown.set( instance.id, instance )
-    } else if ( this.selected.has( instance.id ) ) {
-      this.selected.set( instance.id, instance )
-    }
-  }
-
-  findOrCreateManifestation( vectors )
-  {
-    const created = mesh.createInstance( vectors )
-    const { id } = created
-    const existing = this.shown.get( id ) || this.hidden.get( id ) || this.selected.get( id )
-    if ( existing )
-      return existing.vectors;
-    // TODO avoid creating zero-length struts, to match Java semantics of RMI.findConstruction
-    return vectors
-  }
-
-  showManifestation( vectors )
-  {
-    let instance = mesh.createInstance( vectors )
-    if ( this.shown.has( instance.id ) || this.selected.has( instance.id ) )
-      return // idempotent
-    this.hidden.delete( instance.id )
-    this.shown.set( instance.id, instance )
-  }
-
-  hideManifestation( vectors )
-  {
-    let instance = mesh.createInstance( vectors )
-    instance = this.shown.get( instance.id ) || this.selected.get( instance.id ) || this.hidden.get( instance.id ) || instance
-    this.selected.delete( instance.id ) || this.shown.delete( instance.id )
-    this.hidden.set( instance.id, instance )
-  }
-
-  allIterator()
-  {
-    const shownArray = Array.from( this.shown.values() )
-    const selectedArray = Array.from( this.selected.values() )
-    const hiddenArray = Array.from( this.hidden.values() )
-    const all = [ ...shownArray, ...selectedArray, ...hiddenArray ]
-    return all.map( instance => instance.vectors ).values()
-  }
 }
 
 class Properties
