+++ core/src/jsweet/java/com/vzome/jsweet/JsBall.java
@@ -1,16 +1,10 @@
 package com.vzome.jsweet;
 
-import java.util.Iterator;
-
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-
 import com.vzome.core.algebra.AlgebraicField;
 import com.vzome.core.algebra.AlgebraicVector;
 import com.vzome.core.construction.Construction;
 import com.vzome.core.construction.FreePoint;
 import com.vzome.core.model.Connector;
-import com.vzome.core.model.Group;
 
 import def.js.Object;
 
@@ -33,44 +27,13 @@ public class JsBall extends JsManifestation implements Connector
         return new FreePoint( getLocation() );
     }
 
-    @Override
-    public Group getContainer()
-    {
-        return null;
-    }
-
     
     
     
-    
-    @Override
-    public Iterator<Construction> getConstructions()
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
-    @Override
-    public Element getXml(Document doc)
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
-    @Override
-    public AlgebraicVector getCentroid()
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
-    @Override
-    public void setContainer(Group container)
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
 
     @Override
     public int compareTo( Connector other )
     {
         throw new RuntimeException( "unimplemented" );
     }
-
 }
