+++ core/src/jsweet/java/com/vzome/jsweet/JsStrut.java
@@ -1,16 +1,10 @@
 package com.vzome.jsweet;
 
-import java.util.Iterator;
-
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-
 import com.vzome.core.algebra.AlgebraicField;
 import com.vzome.core.algebra.AlgebraicVector;
 import com.vzome.core.construction.Construction;
 import com.vzome.core.construction.FreePoint;
 import com.vzome.core.construction.SegmentJoiningPoints;
-import com.vzome.core.model.Group;
 import com.vzome.core.model.Strut;
 
 import def.js.Object;
@@ -32,12 +26,6 @@ public class JsStrut extends JsManifestation implements Strut
     public Construction toConstruction()
     {
         return new SegmentJoiningPoints( new FreePoint( getLocation() ), new FreePoint( getEnd() ) );
-    }
-
-    @Override
-    public Group getContainer()
-    {
-        return null;
     }
 
     @Override
@@ -67,30 +55,6 @@ public class JsStrut extends JsManifestation implements Strut
     
     
     
-    @Override
-    public Iterator<Construction> getConstructions()
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
-    @Override
-    public Element getXml(Document doc)
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
-    @Override
-    public AlgebraicVector getCentroid()
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
-    @Override
-    public void setContainer(Group container)
-    {
-        throw new RuntimeException( "unimplemented" );
-    }
-
     @Override
     public int compareTo(Strut other)
     {
