+++ core/src/main/java/com/vzome/core/model/Group.java
@@ -5,6 +5,7 @@ package com.vzome.core.model;
 
 import java.util.ArrayList;
 
+@SuppressWarnings("serial")
 public class Group extends ArrayList<GroupElement> implements GroupElement
 {
     private Group mContainer;
